Move extractInterpreterInfo().
