+++ src/client/common/process/pythonDaemon.ts
@@ -5,12 +5,12 @@
 
 import { ChildProcess } from 'child_process';
 import { MessageConnection, RequestType, RequestType0 } from 'vscode-jsonrpc';
+import { extractInterpreterInfo } from '../../pythonEnvironments/discovery/info';
 import { InterpreterInformation } from '../../pythonEnvironments/discovery/types';
 import { PythonExecInfo } from '../../pythonEnvironments/types';
 import { traceWarning } from '../logger';
 import { BasePythonDaemon } from './baseDaemon';
 import { PythonEnvInfo } from './internal/scripts';
-import { extractInterpreterInfo } from './pythonEnvironment';
 import {
     IPythonDaemonExecutionService,
     IPythonExecutionService,
